+++ rastervision_pytorch_learner/rastervision/pytorch_learner/classification_learner_config.py
@@ -24,10 +24,14 @@ class ClassificationLearnerConfig(LearnerConfig):
     data: ClassificationDataConfig
     model: ClassificationModelConfig
 
-    def build(self, tmp_dir, model_path=None):
+    def build(self, tmp_dir, model_path=None, model_def_path=None):
         from rastervision.pytorch_learner.classification_learner import (
             ClassificationLearner)
-        return ClassificationLearner(self, tmp_dir, model_path=model_path)
+        return ClassificationLearner(
+            self,
+            tmp_dir=tmp_dir,
+            model_path=model_path,
+            model_def_path=model_def_path)
 
     def validate_config(self):
         super().validate_config()