+++ rastervision_pytorch_learner/rastervision/pytorch_learner/learner_config.py
@@ -1,8 +1,8 @@
 from os.path import join
 from enum import Enum
 
-from typing import List, Optional, Union, TYPE_CHECKING
-from pydantic import PositiveFloat, PositiveInt
+from typing import (List, Optional, Union, TYPE_CHECKING)
+from pydantic import PositiveFloat, PositiveInt, constr
 
 from rastervision.pipeline.config import (Config, register_config, ConfigError,
                                           Field)
@@ -97,6 +97,40 @@ class Backbone(Enum):
         return mapping[x]
 
 
+NonEmptyStr = constr(strip_whitespace=True, min_length=1)
+
+
+@register_config('external-module')
+class ExternalModuleConfig(Config):
+    uri: Optional[NonEmptyStr] = Field(
+        None,
+        description=('Local uri of a zip file, or local uri of a directory,'
+                     'or remote uri of zip file.'))
+    github_repo: Optional[constr(
+        strip_whitespace=True, regex=r'.+/.+')] = Field(
+            None, description='<repo-owner>/<repo-name>[:tag]')
+    name: Optional[NonEmptyStr] = Field(
+        None,
+        description=
+        'Name of the folder in which to extract/copy the definition files.')
+    entrypoint: NonEmptyStr = Field(
+        ...,
+        description=('Name of a callable present in hubconf.py. '
+                     'See docs for torch.hub for details.'))
+    entrypoint_args: list = Field(
+        [], description='Args to pass to the entrypoint.')
+    entrypoint_kwargs: dict = Field(
+        {}, description='Keyword args to pass to the entrypoint.')
+    force_reload: bool = Field(
+        False, description='Force reload of the module definition.')
+
+    def validate_config(self):
+        has_uri = self.uri is not None
+        has_repo = self.github_repo is not None
+        if has_uri == has_repo:
+            raise ConfigError('Must specify one of github_repo and uri.')
+
+
 def model_config_upgrader(cfg_dict, version):
     if version == 0:
         cfg_dict['backbone'] = Backbone.int_to_str(cfg_dict['backbone'])
@@ -118,6 +152,10 @@ class ModelConfig(Config):
         None,
         description=('URI of PyTorch model weights used to initialize model. '
                      'If set, this supercedes the pretrained option.'))
+    external_def: Optional[ExternalModuleConfig] = Field(
+        None,
+        description='If specified, the model will be built from the '
+        'definition from this external source, using Torch Hub.')
 
     def update(self, learner: Optional['LearnerConfig'] = None):
         pass
@@ -281,14 +319,18 @@ class LearnerConfig(Config):
             raise ConfigError(
                 'Cannot run_tensorboard if log_tensorboard is False')
 
-    def build(self, tmp_dir: str,
-              model_path: Optional[str] = None) -> 'Learner':
+    def build(self,
+              tmp_dir: str,
+              model_path: Optional[str] = None,
+              model_def_path: Optional[str] = None) -> 'Learner':
         """Returns a Learner instantiated using this Config.
 
         Args:
             tmp_dir: root of temp dirs
             model_path: local path to model weights. If this is passed, the Learner
                 is assumed to be used to make predictions and not train a model.
+            model_def_path: a local path to a directory with a hubconf.py. If
+                provided, the model definition is imported from here.
         """
         raise NotImplementedError()